submit button on edit card working correctly

+++ src/components/PlantCard.js
@@ -13,49 +13,51 @@ class PlantCard extends React.Component {
     }
   }
   deleteThething = () => {
-    this.props.deletePlant(this.props.plant)
+    const result = window.confirm('Are you sure you want to delete this item?');
+    if (result) {
+      return this.props.deletePlant(this.props.plant)
+    }
   }
   switchEdit = () => {
     this.setState({ editing: !this.state.editing });
   }
   render() {
-    const { plant } = this.props;
     return this.state.editing ? (
     <div className="ui raised fluid card">
-      <PlantForm plant={plant} editing={true} />
-      <button className="ui blue button right floated" onClick={this.switchEdit}>Close</button>
+      <div className="content">
+        <PlantForm plant={this.props.plant} edit={true} switchEdit={this.switchEdit} />
+        <button className="ui grey small button right floated" onClick={this.switchEdit}>back</button>
+      </div>
     </div>
     ) : (
       <div className="ui raised fluid plant card">
         <div className="content">
           <div className="ui medium image left floated">
-            <img src={plant.imageUrl} alt={plant.commonName} />
+            <img src={this.props.plant.imageUrl} alt={this.props.plant.commonName} />
           </div>
           <div className="header">
-            {plant.commonName}
+            <span>{this.props.plant.commonName}</span>
             <button className="ui red button right floated right attached" onClick={this.deleteThething}>DELETE</button>
             <button className="ui teal button right floated left attached" onClick={this.switchEdit}>EDIT</button>
           </div>
           <div className="meta">
-            <span>species: {plant.species}</span>
+            <span>species: {this.props.plant.species}</span>
             <br />
-            <span>variety: {plant.variety}</span>
+            <span>variety: {this.props.plant.variety}</span>
             <br />
-            <span>Sun Requirement: {plant.sunReq}</span>
+            <span>Sun Requirement: {this.props.plant.sunReq}</span>
           </div>
-          <br />
-
-          <div className="left floated"><h4>Notes:</h4>{plant.notes}</div>
+          <div className="left floated"><h4>Notes:</h4>{this.props.plant.notes}</div>
           <div className="description">
             <div className="ui grid">
-              <div className="one wide column"><h4>Height:</h4> {plant.height} inches</div>
-              <div className="two wide column"><h4>Planting Depth:</h4> {plant.plantingDepth} inches</div>
-              <div className="two wide column"><h4>Plant Spacing:</h4> {plant.plantSpacing} inches</div>
-              <div className="two wide column"><h4>Color:</h4> {plant.color}</div>
-              <div className="one wide column"><h4>Number:</h4> {plant.number}</div>
-              <div className="two wide column"><h4>Days to bloom:</h4> {plant.daysToBloom}</div>
-              <div className="two wide column"><h4>Germination Temp:</h4> {plant.germinationTemp}</div>
-              <div className="two wide column"><h4>Seeding Date:</h4> {plant.seedingDate}</div>
+              <div className="one wide column"><h4>Height:</h4> {this.props.plant.height} inches</div>
+              <div className="two wide column"><h4>Planting Depth:</h4> {this.props.plant.plantingDepth} inches</div>
+              <div className="two wide column"><h4>Plant Spacing:</h4> {this.props.plant.plantSpacing} inches</div>
+              <div className="two wide column"><h4>Color:</h4> {this.props.plant.color}</div>
+              <div className="one wide column"><h4>Number:</h4> {this.props.plant.number}</div>
+              <div className="two wide column"><h4>Days to bloom:</h4> {this.props.plant.daysToBloom}</div>
+              <div className="two wide column"><h4>Germination Temp:</h4> {this.props.plant.germinationTemp}</div>
+              <div className="two wide column"><h4>Seeding Date:</h4> {this.props.plant.seedingDate.toString()}</div>
             </div>
           </div>
         </div>
